(file) move LoginRequired.tsx to features/member
