+++ src/router/index.tsx
@@ -4,7 +4,7 @@ import Main from '@/features/Map/Main';
 import CompanyDetail from '@/features/Company/CompanyDetail';
 import Mypage from '@/features/Member/Mypage';
 import Login from '@/features/Member/Login';
-import LoginRequired from '@/pages/LoginRequired';
+import LoginRequired from '@/features/Member/LoginRequired';
 import ToBeContinued from '@/pages/ToBeContinued';
 import MainLayout from '@/layout/MainLayout';
 import NotFound from '@/pages/NotFound';
